requirements.txt
Flask
PyPDF2
openai
python-dotenv
Werkzeug